reorganize folder structure within components
