+++ client/src/ui/components/profile/ReviewCard.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReviewBox from "./ReviewBox";
-import { Review } from "../../types/Review";
+import { Review } from "../../../types/Review";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faStar } from "@fortawesome/free-solid-svg-icons";
