+++ client/src/ui/components/sync/ProfileCard.tsx
