+++ client/src/ui/pages/LivePage.tsx
@@ -1,8 +1,8 @@
 import "../../App.css";
-import Navigation from "../components/Navigation.tsx";
-import GameSelection from "../components/GameSelection.tsx";
-import LiveGame from "../components/LiveGame.tsx";
-import CreateLobby from "../components/CreateLobby.tsx";
+import Navigation from "../components/common/Navigation";
+import GameSelection from "../components/live/GameSelection.tsx";
+import LiveGame from "../components/live/LiveGame.tsx";
+import CreateLobby from "../components/live/CreateLobby.tsx";
 import testLobbies from "../../mockData/testLobbies.json";
 import { useEffect, useState } from "react";
 import windowResize from "../utils/WindowResize.tsx";
