+++ client/src/ui/pages/ProfilePage.tsx
@@ -1,12 +1,12 @@
 import "../../App.css";
-import Navigation from "../components/Navigation.tsx";
-import Profile from "../components/Profile.tsx";
-import ProfileCarousel from "../components/ProfileCarousel.tsx";
+import Navigation from "../components/common/Navigation.tsx";
+import Profile from "../components/profile/Profile.tsx";
+import ProfileCarousel from "../components/profile/ProfileCarousel.tsx";
 import { getData } from "../../api/index.ts";
 import { useEffect } from "react";
 import testUser from "../../mockData/testUser.json";
-import ReviewCard from "../components/ReviewCard.tsx";
-import Personality from "../components/Personality.tsx";
+import ReviewCard from "../components/profile/ReviewCard.tsx";
+import Personality from "../components/profile/Personality.tsx";
 import testReviewList from "../../mockData/testReviewList.json";
 import { Review } from "../../types/Review.tsx";
